+++ scripts/declarations/validate/index.mocha.js
@@ -97,6 +97,22 @@ export default async options => {
           }
         });
 
+        if (service) {
+          service.getTermsTypes()
+            .filter(type => service.terms[type]?.latest)
+            .forEach(type => {
+              const terms = service.getTerms({ type });
+
+              if (terms.hasMultipleSourceDocuments) {
+                it(`does not declare the same source document more than once within "${type}"`, () => {
+                  const duplicateLocations = [...new Set(terms.duplicateSourceDocuments.map(sourceDocument => sourceDocument.location))];
+
+                  expect(duplicateLocations, `The same source document is declared more than once within the "${type}" combine: ${duplicateLocations.join(', ')}`).to.be.empty;
+                });
+              }
+            });
+        }
+
         if (!schemaOnly && service) {
           service.getTermsTypes()
             .filter(termsType => {
@@ -117,14 +133,6 @@ export default async options => {
             .forEach(type => {
               describe(type, () => {
                 const terms = service.getTerms({ type });
-
-                if (terms.hasMultipleSourceDocuments) {
-                  it('does not declare the same source document more than once', () => {
-                    const duplicateLocations = [...new Set(terms.duplicateSourceDocuments.map(sourceDocument => sourceDocument.location))];
-
-                    expect(duplicateLocations, `The same source document is declared more than once within the "${type}" combine: ${duplicateLocations.join(', ')}`).to.be.empty;
-                  });
-                }
 
                 terms.sourceDocuments.forEach(sourceDocument => {
                   let filteredContent;